feat: create order counter component
