--- src/components/order-counter/orderCounter.css
@@ -1,11 +1,0 @@
-section {
-    display: flex;
-   }
-   
-   button {
-    flex: 1;
-   }
-   
-   span {
-    font-size: 20px;
-   }
