+++ src/pages/beverage/beverage-details.scss
@@ -62,6 +62,11 @@
     width: 3.5rem;
 }
 
+.button {
+    border-width: 0;
+    background-color: white;
+}
+
 // .maincard {
 //     display: flex;
 //     flex-direction: row
